增加供货商推送商品的服务

+++ iog-service/src/main/java/com/shangpin/iog/dto/OrderDTO.java
@@ -2,8 +2,8 @@ package com.shangpin.iog.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import sun.security.util.BigInt;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -12,7 +12,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class OrderDTO {
-    private BigInt id;
+    private BigInteger id;
     private String supplierId;
     private String uuId;  //和供货商公用的订单唯一标识
     private String spOrderId;
